+++ backend/controllers/post.js
@@ -1,4 +1,5 @@
 const fs = require('fs')
+const jwt = require('jsonwebtoken');
 const xss = require('xss')
 const db = require("../models")
 const Post = db.posts
@@ -119,6 +120,9 @@ exports.getOnePost = (req, res, next) => {
 
 exports.modifyPost = (req, res, next) => {
   //init info
+  const token = req.headers.authorization.split(' ')[1];
+  const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+  const userId = decodedToken.userId
   const id = req.params.id;
   const post = {};
 
@@ -131,24 +135,38 @@ exports.modifyPost = (req, res, next) => {
     post.image = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   }
 
-  //if image =>
-  if (req.file){
+  //get one post
+  Post.findByPk(id, { include: "user" })
+    .then((oldpost) => {
 
-    //get one post
-    Post.findByPk(id)
-      .then((oldpost) => {
+      //if the creator =>
+      if (userId === oldpost.user.id) {
 
-        //if no content's post edit
-        if(!post.content){
-          post.content = oldpost.content
-        }
+        //if image =>
+        if (req.file){
 
-        //if there is an original image
-        if(oldpost.image){
+          //if no content's post edit
+          if(!post.content){
+            post.content = oldpost.content
+          }
 
-          //delete original image
-          const filename = oldpost.image.split('/images/')[1];
-          fs.unlink(`images/${filename}`, () => {
+          //if there is an original image
+          if(oldpost.image){
+
+            //delete original image
+            const filename = oldpost.image.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+
+              //edit post
+              Post.update(post, { where: { id: id }})
+              .then(data => {
+                res.status(201).json({ message: 'Post modifié !' })
+              })
+              .catch(error => res.status(500).json({ error }));
+            });
+
+          //if no original image
+          }else{
 
             //edit post
             Post.update(post, { where: { id: id }})
@@ -156,62 +174,73 @@ exports.modifyPost = (req, res, next) => {
               res.status(201).json({ message: 'Post modifié !' })
             })
             .catch(error => res.status(500).json({ error }));
-          });
+          }
+        }
 
-        //if no original image
-        }else{
-
-          //edit post
+        //if not image => edit post
+        else{
           Post.update(post, { where: { id: id }})
           .then(data => {
             res.status(201).json({ message: 'Post modifié !' })
           })
           .catch(error => res.status(500).json({ error }));
         }
-      })
-      .catch(
-        (error) => {
-          res.status(400).json({
-            error: error
-          });
-        }
-      ); 
-  }
-  //if not image => edit post
-  else{
-    Post.update(post, { where: { id: id }})
-    .then(data => {
-      res.status(201).json({ message: 'Post modifié !' })
+
+      //if not => error
+      }else{
+        res.status(401).json({
+          error: new Error('Invalid request!')
+        });
+      }
+      
     })
-    .catch(error => res.status(500).json({ error }));
-  }
+    .catch(
+      (error) => {
+        res.status(400).json({
+          error: error
+        });
+      }
+    ); 
 };
 
 exports.deletePost = (req, res, next) => {
+  const token = req.headers.authorization.split(' ')[1];
+  const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+  const roleId = decodedToken.roleId
+  const userId = decodedToken.userId
   const id = req.params.id;
 
   //find one post
-  Post.findByPk(id)
+  Post.findByPk(id, { include: "user" })
   .then((post) => {
 
-    //if there is an original image =>
-    if (post.image) {
+    //if admin or if the creator =>
+    if (roleId === 2 || userId === post.user.id) {
+      //if there is an original image =>
+      if (post.image) {
 
-      //delete original image
-      const filename = post.image.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
+        //delete original image
+        const filename = post.image.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
 
-        //delete post
+          //delete post
+          Post.destroy({ where: { id : id }})
+          .then(() => res.status(200).json({ message: 'Post supprimé !'}))
+          .catch(error => res.status(400).json({ error }));
+        });
+
+      //if no original image => delete post
+      }else{
         Post.destroy({ where: { id : id }})
         .then(() => res.status(200).json({ message: 'Post supprimé !'}))
         .catch(error => res.status(400).json({ error }));
-      });
+      }
 
-    //if no original image => delete post
+    //if not => error
     }else{
-      Post.destroy({ where: { id : id }})
-      .then(() => res.status(200).json({ message: 'Post supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
+      res.status(401).json({
+        error: new Error('Invalid request!')
+      });
     }
   })
   .catch(
